+++ db/migrate/20170104162351_remove_deprecated_options_from_domains.rb
@@ -1,0 +1,5 @@
+class RemoveDeprecatedOptionsFromDomains < ActiveRecord::Migration[5.0]
+  def change
+    remove_column :domains, :deprecated_options, :text
+  end
+end