+++ CHANGELOG.md
@@ -9,6 +9,9 @@
   - Updated to jquery-ui-rails 6.0.1
   - Updated to chosen 1.6.2
 
+### Refactoring
+- Removed deprecated_options from domains
+
 ## 0.46.4 (January 3, 2017)
 
 ### Enhancements